requirements.txt
streamlit>=1.36.0
pandas>=2.1.0
yfinance>=0.2.36
requests>=2.31.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
CoinMarketCapAPI>=0.2
